+++ HardwareRetroAchievements.Core/Console/IConsoleRam.cs
@@ -6,9 +6,9 @@ namespace HardwareRetroAchievements.Core.Console
 {
     public interface IConsoleRam
     {
-        public byte ReadInt8(long address);
-        public ushort ReadInt16(long address);
-        public uint ReadUInt24(long address);
-        public uint ReadUInt32(long address);
+        public byte ReadInt8(int address);
+        public ushort ReadInt16(int address);
+        public uint ReadUInt24(int address);
+        public uint ReadUInt32(int address);
     }
 }
